+++ packages/react/src/components/AppShell/src/components/AppShellHeader/appShellHeader.types.ts
@@ -1,5 +1,7 @@
 import { HTMLAttributes, ReactNode } from "react";
 
+import { BoxProperties } from "@nimbus-ds/components";
+
 export interface AppShellHeaderProperties {
   /**
    * Optional content for the left-hand-side slot.
@@ -14,4 +16,5 @@ export interface AppShellHeaderProperties {
 }
 
 export type AppShellHeaderProps = AppShellHeaderProperties &
-  Omit<HTMLAttributes<HTMLElement>, "color">;
+  Omit<HTMLAttributes<HTMLElement>, "color"> &
+  Omit<BoxProperties, "display" | "justifyContent">;