+++ packages/react/src/components/DataList/src/dataList.stories.tsx
@@ -64,5 +64,47 @@ const Template: ComponentStory<typeof DataList.Row> = (args) => {
   );
 };
 
+const NoDividersTemplate: ComponentStory<typeof DataList.Row> = (args) => {
+  const [rows] = useState<RowProps[]>(generateOrders(30));
+
+  return (
+    <DataList bottomDivider={false}>
+      {rows.map((row) => (
+        <DataList.Row {...args} key={row.id} gap="1">
+          <Box display="flex" justifyContent="space-between">
+            <Text fontWeight="medium" color="primary-interactive">
+              #{row.id}
+            </Text>
+            <Text>{formatDate(row.date)}</Text>
+          </Box>
+          <Box display="flex" justifyContent="space-between">
+            <Text>{row.clientName}</Text>
+            <Text>{row.total}</Text>
+          </Box>
+          <Link textDecoration="none">
+            {row.qty} products
+            <Icon color="currentColor" source={<ChevronDownIcon />} />
+          </Link>
+          <Box display="flex" flexWrap="wrap" gap="2" pt="2">
+            <Tag appearance="warning">
+              <Icon color="currentColor" source={<CreditCardIcon />} />
+              Awaiting payment
+            </Tag>
+            <Tag appearance="primary">
+              <Icon color="currentColor" source={<BoxUnpackedIcon />} />
+              Unfulfilled
+            </Tag>
+          </Box>
+        </DataList.Row>
+      ))}
+    </DataList>
+  );
+};
+
 export const base = Template.bind({});
 base.args = {};
+
+export const noDividers = NoDividersTemplate.bind({});
+noDividers.args = {
+  topDivider: false,
+};