+++ packages/react/src/components/DataTable/src/components/DataTableCell/index.ts
@@ -1,0 +1,4 @@
+import { DataTableCell } from "./DataTableCell";
+
+export { DataTableCell } from "./DataTableCell";
+export default DataTableCell;