+++ packages/react/src/components/DataTable/src/components/index.ts
@@ -2,3 +2,4 @@ export * from "./DataTableBulkActions";
 export * from "./DataTableFooter";
 export * from "./DataTableHeader";
 export * from "./DataTableRow";
+export * from "./DataTableCell";