+++ packages/react/src/components/InteractiveList/src/components/InteractiveListToggleItem/interactiveListToggleItem.types.ts
@@ -1,6 +1,7 @@
 import { HTMLAttributes } from "react";
 import { ToggleProperties } from "@nimbus-ds/components";
 import { InteractiveListStructureProperties } from "../InteractiveListStructure/interactiveListStructure.types";
+import { InteractiveListRowProps } from "../InteractiveListRow/interactiveListRow.types";
 
 export interface InteractiveListToggleItemProperties
   extends InteractiveListStructureProperties {
@@ -12,4 +13,5 @@ export interface InteractiveListToggleItemProperties
 
 export type InteractiveListToggleItemProps =
   InteractiveListToggleItemProperties &
-    Omit<HTMLAttributes<HTMLElement>, "color">;
+    Omit<HTMLAttributes<HTMLElement>, "color"> &
+    Pick<InteractiveListRowProps, "topDivider">;