+++ packages/react/src/components/Page/src/components/PageBody/pageBody.types.ts
@@ -1,5 +1,7 @@
 import { ReactNode, HTMLAttributes } from "react";
 
+import { BoxProperties } from "@nimbus-ds/components";
+
 export interface PageBodyProperties {
   /**
    * Content of the page body.
@@ -9,4 +11,5 @@ export interface PageBodyProperties {
 }
 
 export type PageBodyProps = PageBodyProperties &
-  Omit<HTMLAttributes<HTMLElement>, "color">;
+  Omit<HTMLAttributes<HTMLElement>, "color"> &
+  BoxProperties;